requirements.txt
flask
pandas
werkzeug
python-docx
docxtpl
openpyxl
Flask-Login
Flask-SQLAlchemy